requirements.txt
celery==5.2.7
Django==4.2.1
django-celery-beat==2.5.0
django-filter==23.2
djangorestframework==3.14.0
geopy==2.3.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.5
